Save less info, add requirements, split main.py
